requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.6.1
Flask-APScheduler==1.13.1
PyMySQL==1.1.0
redis==5.0.1
python-dotenv==1.0.1
eventlet==0.36.1
requests==2.32.3
simple-websocket==1.1.0
